refactor(core): split god modules — mod.rs and helpers.rs (#118)
