+++ src/memory_core/storage/sqlite/helpers.rs
@@ -1,76 +1,7 @@
-use std::sync::MutexGuard;
-use std::time::Duration;
-
 use super::*;
 
 /// Fallback timestamp used when a row's `created_at` column is missing or unparseable.
 pub(super) const EPOCH_FALLBACK: &str = "1970-01-01T00:00:00.000Z";
-
-/// Returns `true` when a rusqlite error indicates SQLite lock contention
-/// (`SQLITE_BUSY`). Used to decide whether to retry a transaction.
-pub(super) fn is_lock_error(err: &rusqlite::Error) -> bool {
-    matches!(
-        err,
-        rusqlite::Error::SqliteFailure(e, _)
-            if e.code == rusqlite::ffi::ErrorCode::DatabaseBusy
-    )
-}
-
-/// Maximum number of retry attempts for lock contention.
-const RETRY_MAX_ATTEMPTS: u32 = 5;
-
-/// Base delay between retries (doubled each attempt).
-const RETRY_BASE_DELAY_MS: u64 = 10;
-
-/// Retries `f` with exponential backoff when it returns a lock contention error.
-///
-/// This is a **synchronous** function intended to run inside `spawn_blocking`.
-/// Uses `std::thread::sleep` (not tokio) for delays.
-///
-/// - Max attempts: 5
-/// - Backoff: 10ms, 20ms, 40ms, 80ms, 160ms (+ random 0-50% jitter)
-/// - Non-lock errors are returned immediately without retry.
-pub(super) fn retry_on_lock<T, F>(mut f: F) -> std::result::Result<T, rusqlite::Error>
-where
-    F: FnMut() -> std::result::Result<T, rusqlite::Error>,
-{
-    let mut attempt = 0_u32;
-    loop {
-        match f() {
-            Ok(val) => return Ok(val),
-            Err(err) if is_lock_error(&err) && attempt + 1 < RETRY_MAX_ATTEMPTS => {
-                attempt += 1;
-                let base_ms = RETRY_BASE_DELAY_MS * 2_u64.pow(attempt - 1);
-                // Simple jitter: add 0-50% of base_ms using a cheap hash of the attempt
-                // counter and a timestamp nanos component to avoid pulling in a rand dependency.
-                let jitter_ms = {
-                    let nanos = std::time::SystemTime::now()
-                        .duration_since(std::time::UNIX_EPOCH)
-                        .map(|d| d.subsec_nanos() as u64)
-                        .unwrap_or(0);
-                    let seed = nanos.wrapping_mul(u64::from(attempt).wrapping_add(7));
-                    seed % (base_ms / 2 + 1)
-                };
-                let delay = Duration::from_millis(base_ms + jitter_ms);
-                tracing::debug!(
-                    attempt,
-                    delay_ms = delay.as_millis(),
-                    "sqlite lock contention, retrying"
-                );
-                std::thread::sleep(delay);
-            }
-            Err(err) => {
-                if is_lock_error(&err) {
-                    tracing::warn!(
-                        attempts = RETRY_MAX_ATTEMPTS,
-                        "sqlite lock contention persisted after all retries"
-                    );
-                }
-                return Err(err);
-            }
-        }
-    }
-}
 
 /// Computes a u64 hash key for the query result cache.
 pub(super) fn query_cache_key(query: &str, limit: usize, opts: &super::SearchOptions) -> u64 {
@@ -97,383 +28,6 @@ pub(super) fn query_cache_key(query: &str, limit: usize, opts: &super::SearchOpt
     opts.event_before.hash(&mut hasher);
     opts.explain.hash(&mut hasher);
     hasher.finish()
-}
-
-/// Returns `true` when the query looks like a keyword/identifier lookup
-/// that should skip ONNX embedding and vector search (FTS5 only).
-///
-/// Heuristics:
-/// - Contains backtick-wrapped code (`` `...` ``)
-/// - Looks like a file path (contains `/` with no spaces)
-/// - Is a CamelCase identifier (2+ uppercase letters, no spaces)
-/// - Is a snake_case identifier (contains `_`, no spaces, all lowercase)
-pub(super) fn is_keyword_query(query: &str) -> bool {
-    let trimmed = query.trim();
-    if trimmed.is_empty() {
-        return false;
-    }
-
-    // Contains backtick-wrapped code
-    if trimmed.matches('`').count() >= 2 {
-        return true;
-    }
-
-    // Looks like a file path (contains `/` with no spaces)
-    if !trimmed.contains(' ') && trimmed.contains('/') {
-        return true;
-    }
-
-    // Is a CamelCase identifier (2+ uppercase letters, no spaces, alphanumeric only)
-    if !trimmed.contains(' ') {
-        let upper_count = trimmed.chars().filter(|c| c.is_uppercase()).count();
-        if upper_count >= 2 && trimmed.chars().all(|c| c.is_alphanumeric()) {
-            return true;
-        }
-    }
-
-    // Is a snake_case identifier (contains `_`, no spaces, all lowercase alphanumeric + `_`)
-    if !trimmed.contains(' ')
-        && trimmed.contains('_')
-        && trimmed
-            .chars()
-            .all(|c| c.is_lowercase() || c.is_ascii_digit() || c == '_')
-    {
-        return true;
-    }
-
-    false
-}
-
-/// Broad intent categories for incoming queries.
-/// Used to adjust scoring multipliers so that, e.g., factual lookups
-/// favour FTS/BM25 while conceptual queries lean on vector similarity.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum QueryIntent {
-    /// Code identifier / file-path lookup — FTS only, skip embedding.
-    Keyword,
-    /// Factual "who/what/when/where" lookups — precise term matching matters.
-    Factual,
-    /// Conceptual "why/how/explain" queries — semantic similarity matters.
-    Conceptual,
-    /// Everything else — balanced scoring.
-    General,
-}
-
-/// Per-intent multipliers applied on top of the base `ScoringParams`.
-///
-/// Each field is a multiplicative factor:
-///   effective_value = base_value × multiplier
-///
-/// `top_k_mult` scales the internal candidate oversampling (not the final
-/// result limit) so that precision-oriented intents can oversample less
-/// while recall-oriented intents pull in more candidates.
-#[derive(Debug, Clone, Copy)]
-pub(super) struct IntentProfile {
-    pub vec_weight_mult: f64,
-    pub fts_weight_mult: f64,
-    pub word_overlap_mult: f64,
-    pub top_k_mult: f64,
-    #[allow(dead_code)] // Planned for future use; currently tested for intent profiling.
-    pub suggested_limit_mult: f64,
-}
-
-impl IntentProfile {
-    /// Returns the tuned profile for the given intent.
-    pub fn for_intent(intent: QueryIntent) -> Self {
-        match intent {
-            QueryIntent::Keyword => Self {
-                vec_weight_mult: 0.0, // no vector search
-                fts_weight_mult: 1.0,
-                word_overlap_mult: 1.3,
-                top_k_mult: 1.0,
-                suggested_limit_mult: 1.0,
-            },
-            QueryIntent::Factual => Self {
-                vec_weight_mult: 1.0,
-                fts_weight_mult: 1.1,
-                word_overlap_mult: 1.15,
-                top_k_mult: 1.0,
-                suggested_limit_mult: 1.0,
-            },
-            QueryIntent::Conceptual => Self {
-                vec_weight_mult: 1.5,
-                fts_weight_mult: 0.85,
-                word_overlap_mult: 0.7,
-                top_k_mult: 1.3,
-                suggested_limit_mult: 1.3,
-            },
-            QueryIntent::General => Self {
-                vec_weight_mult: 1.0,
-                fts_weight_mult: 1.0,
-                word_overlap_mult: 1.0,
-                top_k_mult: 1.0,
-                suggested_limit_mult: 1.0,
-            },
-        }
-    }
-}
-
-/// Classify a query into a `QueryIntent` category.
-///
-/// Uses lightweight heuristics (no model inference):
-/// 1. `Keyword` — code identifiers, file paths (delegates to `is_keyword_query`).
-/// 2. `Factual` — starts with who/what/when/where/which, or asks for
-///    names/numbers/dates. These benefit from exact term matching.
-/// 3. `Conceptual` — starts with why/how/explain/describe, or contains
-///    phrases like "relationship between", "difference between". These
-///    benefit from semantic similarity.
-/// 4. `General` — everything else.
-pub(super) fn classify_query_intent(query: &str) -> QueryIntent {
-    if is_keyword_query(query) {
-        return QueryIntent::Keyword;
-    }
-
-    let lower = query.trim().to_lowercase();
-    if lower.is_empty() {
-        return QueryIntent::General;
-    }
-
-    // Split on first word for prefix matching.
-    let first_word = lower.split_whitespace().next().unwrap_or("");
-
-    // Factual "how many/much/old/long/often" — must be checked before the generic
-    // "how" → Conceptual rule so that "How many children …" is Factual.
-    if lower.starts_with("how many")
-        || lower.starts_with("how much")
-        || lower.starts_with("how old")
-        || lower.starts_with("how long")
-        || lower.starts_with("how often")
-    {
-        return QueryIntent::Factual;
-    }
-
-    // Conceptual signals.
-    if matches!(
-        first_word,
-        "why" | "how" | "explain" | "describe" | "elaborate"
-    ) || lower.starts_with("what is the relationship")
-        || lower.starts_with("what is the difference")
-        || lower.contains("difference between")
-        || lower.contains("relationship between")
-        || lower.contains("compared to")
-        || lower.contains("pros and cons")
-    {
-        return QueryIntent::Conceptual;
-    }
-
-    // Factual signals.
-    if matches!(
-        first_word,
-        "who" | "what" | "when" | "where" | "which" | "name" | "list" | "find"
-    ) || lower.contains("date of")
-        || lower.contains("name of")
-        || lower.contains("number of")
-    {
-        return QueryIntent::Factual;
-    }
-
-    QueryIntent::General
-}
-
-/// Detects whether a query warrants additional retrieval candidates.
-///
-/// Returns a multiplier for the candidate limit:
-/// - 2.0x for multi-hop indicators ("and" combined with relationship/connect/between/both)
-/// - 1.5x for broad temporal queries ("last month/year", "this month", "over the past")
-/// - 1.0x otherwise (no adjustment)
-pub(super) fn detect_dynamic_limit_mult(query: &str) -> f64 {
-    let lower = query.trim().to_lowercase();
-
-    // Multi-hop: "and" + relationship words
-    let has_and = lower.contains(" and ");
-    let multi_hop_indicators = [
-        "relationship",
-        "connect",
-        "between",
-        "both",
-        "related",
-        "sister",
-        "brother",
-        "friend",
-    ];
-    if has_and && multi_hop_indicators.iter().any(|w| lower.contains(w)) {
-        return 2.0;
-    }
-
-    // Broad temporal patterns
-    let temporal_patterns = [
-        "last month",
-        "last year",
-        "this month",
-        "this year",
-        "over the past",
-        "in the past",
-        "recent months",
-        "past year",
-        "past month",
-        "few months",
-    ];
-    if temporal_patterns.iter().any(|p| lower.contains(p)) {
-        return 1.5;
-    }
-
-    1.0
-}
-
-/// Default number of reader connections in the pool for file-backed databases.
-const DEFAULT_READER_COUNT: usize = 4;
-
-/// Default number of writes between WAL checkpoints.
-const WAL_CHECKPOINT_INTERVAL: u64 = 10;
-
-/// Connection pool providing one writer connection and N reader connections.
-///
-/// In WAL mode, SQLite allows concurrent readers while a single writer holds
-/// the write lock. `rusqlite::Connection` is `!Sync`, so each connection is
-/// wrapped in its own `Mutex`.
-///
-/// For in-memory databases (`:memory:` or `file::memory:`) that cannot share
-/// state across connections, the pool falls back to single-connection mode
-/// where all operations use the writer.
-#[derive(Debug)]
-pub(super) struct ConnPool {
-    writer: Mutex<Connection>,
-    readers: Vec<Mutex<Connection>>,
-    /// Round-robin counter for reader selection.
-    reader_idx: std::sync::atomic::AtomicUsize,
-    /// Monotonic write counter for periodic WAL checkpoints.
-    write_count: std::sync::atomic::AtomicU64,
-    /// Whether this pool is file-backed (WAL checkpoints only apply to files).
-    is_file_backed: bool,
-}
-
-impl ConnPool {
-    /// Opens a file-backed connection pool: one writer + N readers.
-    ///
-    /// All connections share the same SQLite database file and run in WAL mode.
-    /// Performs a startup WAL checkpoint (TRUNCATE) to reclaim any stale WAL
-    /// from previous runs.
-    pub(super) fn open_file(path: &Path, embedding_dim: usize) -> Result<Self> {
-        #[cfg(feature = "sqlite-vec")]
-        super::ensure_vec_extension_registered();
-
-        super::initialize_parent_dir(path)?;
-
-        let writer = open_connection(path)?;
-        // Writer sets WAL mode; readers inherit it from the file.
-        initialize_schema(&writer, embedding_dim)?;
-
-        // Startup checkpoint: safe because we're the only writer at init time.
-        if let Err(e) = writer.execute_batch("PRAGMA wal_checkpoint(TRUNCATE);") {
-            tracing::debug!("startup WAL checkpoint skipped: {e}");
-        }
-
-        let mut readers = Vec::with_capacity(DEFAULT_READER_COUNT);
-        for _ in 0..DEFAULT_READER_COUNT {
-            let reader = open_connection(path)?;
-            configure_reader(&reader)?;
-            readers.push(Mutex::new(reader));
-        }
-
-        Ok(Self {
-            writer: Mutex::new(writer),
-            readers,
-            reader_idx: std::sync::atomic::AtomicUsize::new(0),
-            write_count: std::sync::atomic::AtomicU64::new(0),
-            is_file_backed: true,
-        })
-    }
-
-    /// Opens an in-memory pool (single connection, no reader pool).
-    ///
-    /// Used for tests and non-file-backed scenarios where multiple connections
-    /// cannot share state.
-    pub(super) fn open_in_memory(embedding_dim: usize) -> Result<Self> {
-        #[cfg(feature = "sqlite-vec")]
-        super::ensure_vec_extension_registered();
-
-        let conn = Connection::open_in_memory().context("failed to open in-memory sqlite")?;
-        initialize_schema(&conn, embedding_dim)?;
-
-        Ok(Self {
-            writer: Mutex::new(conn),
-            readers: Vec::new(),
-            reader_idx: std::sync::atomic::AtomicUsize::new(0),
-            write_count: std::sync::atomic::AtomicU64::new(0),
-            is_file_backed: false,
-        })
-    }
-
-    /// Acquires the writer connection.
-    pub(super) fn writer(&self) -> Result<MutexGuard<'_, Connection>> {
-        self.writer
-            .lock()
-            .map_err(|_| anyhow!("sqlite writer mutex poisoned"))
-    }
-
-    /// Returns `true` when the pool has dedicated reader connections.
-    pub(super) fn has_readers(&self) -> bool {
-        !self.readers.is_empty()
-    }
-
-    /// Increments the write counter and runs a passive WAL checkpoint every
-    /// [`WAL_CHECKPOINT_INTERVAL`] writes. Call after each write transaction
-    /// commits. No-op for in-memory databases.
-    pub(super) fn note_write(&self) {
-        if !self.is_file_backed {
-            return;
-        }
-        let count = self
-            .write_count
-            .fetch_add(1, std::sync::atomic::Ordering::Relaxed)
-            + 1;
-        if !count.is_multiple_of(WAL_CHECKPOINT_INTERVAL) {
-            return;
-        }
-        if let Ok(conn) = self.writer() {
-            match conn.execute_batch("PRAGMA wal_checkpoint(PASSIVE);") {
-                Ok(()) => tracing::debug!(writes = count, "WAL passive checkpoint completed"),
-                Err(e) => tracing::debug!(writes = count, "WAL passive checkpoint failed: {e}"),
-            }
-        }
-    }
-
-    /// Acquires a reader connection via round-robin. Falls back to the writer
-    /// when no dedicated readers exist (in-memory mode).
-    pub(super) fn reader(&self) -> Result<MutexGuard<'_, Connection>> {
-        if self.readers.is_empty() {
-            return self.writer();
-        }
-        let idx = self
-            .reader_idx
-            .fetch_add(1, std::sync::atomic::Ordering::Relaxed)
-            % self.readers.len();
-        self.readers[idx]
-            .lock()
-            .map_err(|_| anyhow!("sqlite reader mutex poisoned"))
-    }
-}
-
-/// Opens a single SQLite connection with standard PRAGMAs.
-fn open_connection(path: &Path) -> Result<Connection> {
-    let conn = Connection::open(path)
-        .with_context(|| format!("failed to open sqlite database at {}", path.display()))?;
-    configure_reader(&conn)?;
-    Ok(conn)
-}
-
-/// Applies read-oriented PRAGMAs shared by all connections (writer and readers).
-fn configure_reader(conn: &Connection) -> Result<()> {
-    conn.execute_batch(
-        "PRAGMA journal_mode=WAL;\
-         PRAGMA busy_timeout=5000;\
-         PRAGMA cache_size=-16000;\
-         PRAGMA mmap_size=33554432;\
-         PRAGMA synchronous=NORMAL;\
-         PRAGMA temp_store=MEMORY;",
-    )
-    .context("failed to set connection PRAGMAs")?;
-    Ok(())
 }
 
 pub(super) fn canonicalize(content: &str) -> String {
@@ -990,63 +544,6 @@ pub(super) fn build_fts5_query(input: &str) -> String {
 }
 
 /// Encodes a slice of f32 values as little-endian bytes.
-pub(super) fn encode_embedding(v: &[f32]) -> Vec<u8> {
-    let mut buf = Vec::with_capacity(v.len() * 4);
-    for &val in v {
-        buf.extend_from_slice(&val.to_le_bytes());
-    }
-    buf
-}
-
-/// Decodes an embedding BLOB. Tries binary (little-endian f32) first,
-/// falls back to JSON for backwards compatibility with existing data.
-pub(super) fn decode_embedding(blob: &[u8]) -> Result<Vec<f32>> {
-    if blob.is_empty() {
-        return Ok(Vec::new());
-    }
-    // Binary format: length must be a multiple of 4
-    if blob.len().is_multiple_of(4) {
-        // Quick heuristic: JSON always starts with '[' (0x5B)
-        if blob[0] != b'[' {
-            return Ok(decode_binary_embedding(blob));
-        }
-        // First byte is '[' — could be JSON or binary coincidence.
-        // Try JSON first (backwards compat), fall back to binary.
-        if let Ok(v) = serde_json::from_slice::<Vec<f32>>(blob) {
-            return Ok(v);
-        }
-        return Ok(decode_binary_embedding(blob));
-    }
-    // Not a multiple of 4 — must be JSON
-    serde_json::from_slice(blob).context("failed to decode embedding (neither binary nor JSON)")
-}
-
-fn decode_binary_embedding(blob: &[u8]) -> Vec<f32> {
-    blob.chunks_exact(4)
-        .map(|chunk| {
-            let mut bytes = [0_u8; 4];
-            bytes.copy_from_slice(chunk);
-            f32::from_le_bytes(bytes)
-        })
-        .collect()
-}
-
-/// Dot product of two vectors. Equivalent to cosine similarity when inputs are L2-normalized.
-pub(crate) fn dot_product(a: &[f32], b: &[f32]) -> f32 {
-    if a.len() != b.len() || a.is_empty() {
-        return 0.0;
-    }
-    debug_assert!(
-        (a.iter().map(|v| v * v).sum::<f32>().sqrt() - 1.0).abs() < 0.01,
-        "input a is not L2-normalized"
-    );
-    debug_assert!(
-        (b.iter().map(|v| v * v).sum::<f32>().sqrt() - 1.0).abs() < 0.01,
-        "input b is not L2-normalized"
-    );
-    a.iter().zip(b.iter()).map(|(x, y)| x * y).sum()
-}
-
 pub(super) fn resolve_priority(event_type: Option<&EventType>, priority: Option<i64>) -> u8 {
     if let Some(value) = priority
         && (1..=5).contains(&value)
@@ -1483,658 +980,6 @@ pub(super) fn hydrate_memories_by_ids(
     Ok(hydrated)
 }
 
-/// Result of expanding temporal references from a query string.
-#[derive(Debug, Default)]
-pub(super) struct TemporalExpansion {
-    /// The query with temporal phrases stripped out.
-    pub cleaned_query: String,
-    /// ISO 8601 lower bound inferred from the temporal phrase.
-    pub event_after: Option<String>,
-    /// ISO 8601 upper bound inferred from the temporal phrase.
-    pub event_before: Option<String>,
-}
-
-/// Detects temporal references in a query and converts them to date filters.
-///
-/// Recognized patterns:
-/// - "today", "yesterday"
-/// - "this week", "last week", "this month", "last month"
-/// - "N days ago", "N weeks ago", "N months ago"
-/// - "last N days", "last N weeks", "last N months", "past N days/weeks/months"
-///
-/// Returns the cleaned query (temporal phrase removed) and inferred date bounds.
-/// If no temporal phrase is detected, returns the original query unchanged.
-pub(super) fn expand_temporal_query(query: &str, now: &chrono::NaiveDate) -> TemporalExpansion {
-    use chrono::{Datelike, Duration, NaiveDate};
-
-    let lower = query.to_lowercase();
-
-    // Try each pattern; first match wins.
-    let result = if let Some(idx) = lower.find("yesterday") {
-        let cleaned = remove_phrase(query, idx, 9);
-        let date = (*now - Duration::days(1)).format("%Y-%m-%d").to_string();
-        Some((
-            cleaned,
-            Some(date.clone()),
-            Some(format!("{date}T23:59:59")),
-        ))
-    } else if let Some(idx) = lower.find("today") {
-        let cleaned = remove_phrase(query, idx, 5);
-        let date = now.format("%Y-%m-%d").to_string();
-        Some((
-            cleaned,
-            Some(date.clone()),
-            Some(format!("{date}T23:59:59")),
-        ))
-    } else if let Some(re_pos) = lower.find(" ago") {
-        // "N days/weeks/months ago"
-        let before = &lower[..re_pos];
-        let all_words: Vec<&str> = before.split_whitespace().collect();
-        if all_words.len() >= 2 {
-            let unit = all_words[all_words.len() - 1];
-            let num_str = all_words[all_words.len() - 2];
-            num_str.parse::<i64>().ok().and_then(|n| {
-                let (after, before_date) = compute_ago(now, n, unit)?;
-                let phrase = format!("{num_str} {unit} ago");
-                let phrase_start = lower.find(&phrase)?;
-                let cleaned = remove_phrase(query, phrase_start, phrase.len());
-                Some((cleaned, Some(after), Some(before_date)))
-            })
-        } else {
-            None
-        }
-    } else if let Some(m) = try_prefix_n_unit(&lower, query, "last ", now) {
-        Some(m)
-    } else if let Some(m) = try_prefix_n_unit(&lower, query, "past ", now) {
-        Some(m)
-    } else if let Some(idx) = lower.find("last month") {
-        let cleaned = remove_phrase(query, idx, 10);
-        NaiveDate::from_ymd_opt(now.year(), now.month(), 1).and_then(|first_this| {
-            let last_day_prev = first_this - Duration::days(1);
-            let first_prev =
-                NaiveDate::from_ymd_opt(last_day_prev.year(), last_day_prev.month(), 1)?;
-            Some((
-                cleaned,
-                Some(first_prev.format("%Y-%m-%d").to_string()),
-                Some(format!("{}T23:59:59", last_day_prev.format("%Y-%m-%d"))),
-            ))
-        })
-    } else if let Some(idx) = lower.find("last week") {
-        let cleaned = remove_phrase(query, idx, 9);
-        let weekday = now.weekday().num_days_from_monday();
-        let this_monday = *now - Duration::days(weekday as i64);
-        let last_monday = this_monday - Duration::days(7);
-        let last_sunday = this_monday - Duration::days(1);
-        Some((
-            cleaned,
-            Some(last_monday.format("%Y-%m-%d").to_string()),
-            Some(format!("{}T23:59:59", last_sunday.format("%Y-%m-%d"))),
-        ))
-    } else if let Some(idx) = lower.find("this week") {
-        let cleaned = remove_phrase(query, idx, 9);
-        let weekday = now.weekday().num_days_from_monday();
-        let start = *now - Duration::days(weekday as i64);
-        Some((cleaned, Some(start.format("%Y-%m-%d").to_string()), None))
-    } else if let Some(idx) = lower.find("this month") {
-        let cleaned = remove_phrase(query, idx, 10);
-        NaiveDate::from_ymd_opt(now.year(), now.month(), 1)
-            .map(|start| (cleaned, Some(start.format("%Y-%m-%d").to_string()), None))
-    } else {
-        None
-    };
-
-    match result {
-        Some((cleaned, after, before)) => {
-            let trimmed = cleaned.split_whitespace().collect::<Vec<_>>().join(" ");
-            if trimmed.is_empty() {
-                TemporalExpansion {
-                    cleaned_query: query.to_string(),
-                    event_after: after,
-                    event_before: before,
-                }
-            } else {
-                TemporalExpansion {
-                    cleaned_query: trimmed,
-                    event_after: after,
-                    event_before: before,
-                }
-            }
-        }
-        None => TemporalExpansion {
-            cleaned_query: query.to_string(),
-            event_after: None,
-            event_before: None,
-        },
-    }
-}
-
-/// Tries to match "last N unit" or "past N unit" at `prefix` position.
-fn try_prefix_n_unit(
-    lower: &str,
-    query: &str,
-    prefix: &str,
-    now: &chrono::NaiveDate,
-) -> Option<(String, Option<String>, Option<String>)> {
-    let idx = lower.find(prefix)?;
-    let rest = &lower[idx + prefix.len()..];
-    let tokens: Vec<&str> = rest.split_whitespace().take(2).collect();
-    if tokens.len() < 2 {
-        return None;
-    }
-    let n: i64 = tokens[0].parse().ok()?;
-    let unit = tokens[1];
-    let (after, before_date) = compute_ago(now, n, unit)?;
-    let phrase_len = prefix.len() + tokens[0].len() + 1 + tokens[1].len();
-    let cleaned = remove_phrase(query, idx, phrase_len);
-    Some((cleaned, Some(after), Some(before_date)))
-}
-
-fn remove_phrase(query: &str, start: usize, len: usize) -> String {
-    let mut result = String::with_capacity(query.len());
-    result.push_str(&query[..start]);
-    if start + len < query.len() {
-        result.push_str(&query[start + len..]);
-    }
-    result
-}
-
-fn compute_ago(now: &chrono::NaiveDate, n: i64, unit: &str) -> Option<(String, String)> {
-    use chrono::{Datelike, Duration};
-    let start = match unit.trim_end_matches('s') {
-        "day" => *now - Duration::days(n),
-        "week" => *now - Duration::weeks(n),
-        "month" => {
-            // Approximate: subtract n months
-            let n_i32 = i32::try_from(n).ok()?;
-            #[allow(clippy::cast_sign_loss)]
-            let total_months = now.year() * 12 + now.month() as i32 - 1 - n_i32;
-            let year = total_months / 12;
-            let month = u32::try_from(total_months % 12 + 1).ok()?;
-            let day = now.day().min(28); // safe day
-            chrono::NaiveDate::from_ymd_opt(year, month, day)?
-        }
-        _ => return None,
-    };
-    Some((
-        start.format("%Y-%m-%d").to_string(),
-        format!("{}T23:59:59", now.format("%Y-%m-%d")),
-    ))
-}
-
-/// Stopwords for entity extraction from queries — words that appear capitalized
-/// but are not entities (question words, auxiliaries, months, etc.)
-const ENTITY_STOPWORDS: &[&str] = &[
-    "What",
-    "How",
-    "Why",
-    "When",
-    "Where",
-    "Which",
-    "Who",
-    "Whose",
-    "Whom",
-    "Is",
-    "Are",
-    "Was",
-    "Were",
-    "Do",
-    "Does",
-    "Did",
-    "Has",
-    "Have",
-    "Had",
-    "Will",
-    "Would",
-    "Could",
-    "Should",
-    "Can",
-    "May",
-    "Might",
-    "Must",
-    "Being",
-    "Been",
-    "Having",
-    "The",
-    "This",
-    "That",
-    "These",
-    "Those",
-    "It",
-    "Its",
-    "Yes",
-    "No",
-    "Not",
-    "Very",
-    "Also",
-    "Just",
-    "Even",
-    "Still",
-    "January",
-    "February",
-    "March",
-    "April",
-    // "May" already listed above (line 1686)
-    "June",
-    "July",
-    "August",
-    "September",
-    "October",
-    "November",
-    "December",
-    "Monday",
-    "Tuesday",
-    "Wednesday",
-    "Thursday",
-    "Friday",
-    "Saturday",
-    "Sunday",
-    "National",
-    "American",
-    "European",
-    "Asian",
-    "African",
-    "International",
-    "Global",
-    "Local",
-    "Regional",
-    "Western",
-    "Eastern",
-    "Northern",
-    "Southern",
-    "Answer",
-    "Based",
-    "According",
-    "Since",
-    "Because",
-    "Although",
-    "However",
-    "Therefore",
-    "Furthermore",
-    "Moreover",
-    "Nevertheless",
-    "Likely",
-    "Probably",
-    "Certainly",
-    "Actually",
-    "Recently",
-    "Today",
-    "Yesterday",
-    "Tomorrow",
-];
-
-/// Extract capitalized proper nouns (entity names) from a query string.
-///
-/// Skips sentence-initial words, stopwords, and words < 2 chars.
-/// Also extracts possessives (e.g., "John's" -> "John").
-pub(super) fn extract_query_entities(query: &str) -> Vec<String> {
-    let mut entities = Vec::new();
-    let mut seen = std::collections::HashSet::new();
-    static STOPWORDS_SET: std::sync::LazyLock<std::collections::HashSet<&str>> =
-        std::sync::LazyLock::new(|| ENTITY_STOPWORDS.iter().copied().collect());
-    let stopwords = &*STOPWORDS_SET;
-
-    let words: Vec<&str> = query.split_whitespace().collect();
-
-    for (i, word) in words.iter().enumerate() {
-        let clean: String = word.chars().filter(|c| c.is_alphanumeric()).collect();
-
-        if clean.len() < 2 {
-            continue;
-        }
-        if stopwords.contains(clean.as_str()) {
-            continue;
-        }
-        if i == 0 {
-            continue;
-        }
-        if i > 0 {
-            let prev = words[i - 1];
-            if prev.ends_with('.') || prev.ends_with('?') || prev.ends_with('!') {
-                continue;
-            }
-        }
-        if clean.len() > 1
-            && clean.chars().next().is_some_and(|c| c.is_uppercase())
-            && clean.chars().skip(1).all(|c| c.is_lowercase())
-            && seen.insert(clean.clone())
-        {
-            entities.push(clean);
-        }
-    }
-
-    // Extract possessives: "Name's" -> "Name"
-    for word in &words {
-        if let Some(pos) = word.find("'s") {
-            let name = &word[..pos];
-            if name.len() >= 2
-                && name.chars().next().is_some_and(|c| c.is_uppercase())
-                && name.chars().skip(1).all(|c| c.is_lowercase())
-                && !stopwords.contains(name)
-                && seen.insert(name.to_string())
-            {
-                entities.push(name.to_string());
-            }
-        }
-    }
-
-    entities
-}
-
-/// Extract semantic topic keywords from a query, excluding entities and common words.
-///
-/// Returns up to 5 unique topic words that are 4+ chars and not in the skip list.
-pub(super) fn extract_topic_keywords(query: &str, exclude_entities: &[String]) -> Vec<String> {
-    let exclude_lower: std::collections::HashSet<String> =
-        exclude_entities.iter().map(|e| e.to_lowercase()).collect();
-
-    static SKIP_WORDS: std::sync::LazyLock<std::collections::HashSet<&str>> =
-        std::sync::LazyLock::new(|| {
-            [
-                "what",
-                "when",
-                "where",
-                "which",
-                "who",
-                "whom",
-                "whose",
-                "how",
-                "why",
-                "that",
-                "this",
-                "these",
-                "those",
-                "there",
-                "here",
-                "then",
-                "than",
-                "been",
-                "being",
-                "have",
-                "having",
-                "does",
-                "doing",
-                "done",
-                "will",
-                "would",
-                "could",
-                "should",
-                "shall",
-                "might",
-                "must",
-                "about",
-                "after",
-                "also",
-                "another",
-                "away",
-                "back",
-                "because",
-                "before",
-                "between",
-                "both",
-                "came",
-                "come",
-                "coming",
-                "each",
-                "even",
-                "every",
-                "first",
-                "from",
-                "gets",
-                "give",
-                "given",
-                "goes",
-                "going",
-                "gone",
-                "good",
-                "great",
-                "into",
-                "just",
-                "keep",
-                "kind",
-                "know",
-                "known",
-                "last",
-                "left",
-                "like",
-                "liked",
-                "likely",
-                "long",
-                "look",
-                "made",
-                "make",
-                "making",
-                "many",
-                "more",
-                "most",
-                "much",
-                "need",
-                "never",
-                "next",
-                "once",
-                "only",
-                "other",
-                "over",
-                "part",
-                "past",
-                "people",
-                "perhaps",
-                "place",
-                "point",
-                "probably",
-                "pursue",
-                "quite",
-                "rather",
-                "really",
-                "right",
-                "said",
-                "same",
-                "seem",
-                "show",
-                "side",
-                "since",
-                "some",
-                "something",
-                "sometimes",
-                "still",
-                "such",
-                "sure",
-                "take",
-                "tell",
-                "them",
-                "they",
-                "thing",
-                "things",
-                "think",
-                "time",
-                "told",
-                "took",
-                "turn",
-                "under",
-                "upon",
-                "used",
-                "using",
-                "very",
-                "want",
-                "well",
-                "went",
-                "were",
-                "while",
-                "with",
-                "within",
-                "without",
-                "work",
-                "year",
-                "years",
-                "your",
-                "able",
-                "already",
-                "always",
-                "around",
-                "called",
-                "certain",
-                "during",
-                "either",
-                "else",
-                "enough",
-                "ever",
-                "fact",
-                "find",
-                "found",
-                "hard",
-                "help",
-                "high",
-                "hold",
-                "home",
-                "however",
-                "important",
-                "include",
-                "indeed",
-                "inside",
-                "instead",
-                "itself",
-                "large",
-                "later",
-                "least",
-                "less",
-                "life",
-                "line",
-                "little",
-                "live",
-                "lived",
-                "lives",
-                "living",
-                "longer",
-                "mean",
-                "means",
-                "mention",
-                "mentioned",
-                "name",
-                "near",
-                "number",
-                "often",
-                "open",
-                "order",
-                "others",
-                "outside",
-                "particular",
-                "play",
-                "possible",
-                "prefer",
-                "pretty",
-                "provide",
-                "real",
-                "reason",
-                "regarding",
-                "remember",
-                "result",
-                "seems",
-                "sense",
-                "several",
-                "simply",
-                "small",
-                "sort",
-                "speak",
-                "start",
-                "state",
-                "talk",
-                "together",
-                "true",
-                "type",
-                "types",
-                "usually",
-                "various",
-                "ways",
-            ]
-            .iter()
-            .copied()
-            .collect()
-        });
-    let skip_words = &*SKIP_WORDS;
-
-    let lower = query.to_lowercase();
-
-    let mut topics = Vec::new();
-    let mut seen = std::collections::HashSet::new();
-
-    let mut current_word = String::new();
-    for ch in lower.chars() {
-        if ch.is_ascii_lowercase() {
-            current_word.push(ch);
-        } else {
-            if current_word.len() >= 4
-                && !skip_words.contains(current_word.as_str())
-                && !exclude_lower.contains(&current_word)
-                && seen.insert(current_word.clone())
-            {
-                topics.push(current_word.clone());
-                if topics.len() >= 5 {
-                    return topics;
-                }
-            }
-            current_word.clear();
-        }
-    }
-    if current_word.len() >= 4
-        && !skip_words.contains(current_word.as_str())
-        && !exclude_lower.contains(&current_word)
-        && seen.insert(current_word.clone())
-    {
-        topics.push(current_word);
-    }
-
-    topics
-}
-
-/// Generate decomposed sub-queries from extracted entities and topics.
-///
-/// Returns the original query first, then entity-based and topic-based variations.
-/// Max 2 entities, max 3 topics per entity.
-pub(super) fn generate_sub_queries(
-    query: &str,
-    entities: &[String],
-    topics: &[String],
-) -> Vec<String> {
-    let mut queries = vec![query.to_string()];
-
-    for entity in entities.iter().take(2) {
-        queries.push(entity.clone());
-        for topic in topics.iter().take(3) {
-            queries.push(format!("{entity} {topic}"));
-        }
-        if topics.iter().any(|t| {
-            matches!(
-                t.as_str(),
-                "career" | "jobs" | "work" | "occupation" | "employment"
-            )
-        }) {
-            queries.push(format!("{entity} interests goals plans"));
-        }
-    }
-
-    if entities.is_empty() && !topics.is_empty() {
-        for topic in topics.iter().take(3) {
-            queries.push(topic.clone());
-        }
-    }
-
-    queries
-}
-
-/// Content fingerprint for dedup: first 320 chars, normalized, lowered, ASCII-only.
-pub(super) fn content_fingerprint(content: &str) -> String {
-    let cleaned: String = content
-        .to_lowercase()
-        .chars()
-        .filter(|c| c.is_ascii_alphanumeric() || c.is_ascii_whitespace())
-        .collect();
-    let collapsed = cleaned.split_whitespace().collect::<Vec<_>>().join(" ");
-    if collapsed.len() > 320 {
-        collapsed[..320].to_string()
-    } else {
-        collapsed
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -2200,396 +1045,6 @@ mod tests {
             build_fts5_query("  database   connection   pool  "),
             "\"database\" OR \"connection\" OR \"pool\" \
              OR \"database connection\" OR \"connection pool\""
-        );
-    }
-
-    #[test]
-    fn temporal_today() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("errors today", &now);
-        assert_eq!(exp.cleaned_query, "errors");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-03-08"));
-        assert_eq!(exp.event_before.as_deref(), Some("2026-03-08T23:59:59"));
-    }
-
-    #[test]
-    fn temporal_yesterday() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("what happened yesterday", &now);
-        assert_eq!(exp.cleaned_query, "what happened");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-03-07"));
-        assert_eq!(exp.event_before.as_deref(), Some("2026-03-07T23:59:59"));
-    }
-
-    #[test]
-    fn temporal_n_days_ago() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("decisions 3 days ago", &now);
-        assert_eq!(exp.cleaned_query, "decisions");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-03-05"));
-    }
-
-    #[test]
-    fn temporal_last_n_days() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("bugs last 7 days", &now);
-        assert_eq!(exp.cleaned_query, "bugs");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-03-01"));
-    }
-
-    #[test]
-    fn temporal_this_week() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap(); // Sunday
-        let exp = expand_temporal_query("decisions this week", &now);
-        assert_eq!(exp.cleaned_query, "decisions");
-        // March 8 2026 is Sunday, weekday=6 from Monday. Monday = March 2.
-        assert_eq!(exp.event_after.as_deref(), Some("2026-03-02"));
-        assert!(exp.event_before.is_none()); // open-ended
-    }
-
-    #[test]
-    fn temporal_last_week() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("errors last week", &now);
-        assert_eq!(exp.cleaned_query, "errors");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-02-23"));
-        assert_eq!(exp.event_before.as_deref(), Some("2026-03-01T23:59:59"));
-    }
-
-    #[test]
-    fn temporal_this_month() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 15).unwrap();
-        let exp = expand_temporal_query("lessons this month", &now);
-        assert_eq!(exp.cleaned_query, "lessons");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-03-01"));
-    }
-
-    #[test]
-    fn temporal_last_month() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 15).unwrap();
-        let exp = expand_temporal_query("bugs last month", &now);
-        assert_eq!(exp.cleaned_query, "bugs");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-02-01"));
-        assert_eq!(exp.event_before.as_deref(), Some("2026-02-28T23:59:59"));
-    }
-
-    #[test]
-    fn temporal_no_match() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("database connection pool", &now);
-        assert_eq!(exp.cleaned_query, "database connection pool");
-        assert!(exp.event_after.is_none());
-        assert!(exp.event_before.is_none());
-    }
-
-    #[test]
-    fn temporal_query_only_temporal() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("today", &now);
-        // When only temporal phrase, keep original query for search
-        assert_eq!(exp.cleaned_query, "today");
-        assert!(exp.event_after.is_some());
-    }
-
-    #[test]
-    fn temporal_past_n_weeks() {
-        let now = chrono::NaiveDate::from_ymd_opt(2026, 3, 8).unwrap();
-        let exp = expand_temporal_query("errors past 2 weeks", &now);
-        assert_eq!(exp.cleaned_query, "errors");
-        assert_eq!(exp.event_after.as_deref(), Some("2026-02-22"));
-    }
-
-    // ── is_keyword_query tests ──────────────────────────────────────────
-
-    #[test]
-    fn keyword_backtick_code() {
-        assert!(is_keyword_query("`FooBar`"));
-        assert!(is_keyword_query("find `my_func` usage"));
-    }
-
-    #[test]
-    fn keyword_file_path() {
-        assert!(is_keyword_query("src/main.rs"));
-        assert!(is_keyword_query("./config.toml"));
-        assert!(is_keyword_query("~/projects/foo"));
-        assert!(is_keyword_query("/usr/local/bin"));
-    }
-
-    #[test]
-    fn keyword_camel_case() {
-        assert!(is_keyword_query("SqliteStorage"));
-        assert!(is_keyword_query("McpMemoryServer"));
-        assert!(!is_keyword_query("sqlite"));
-    }
-
-    #[test]
-    fn keyword_snake_case() {
-        assert!(is_keyword_query("query_cache_key"));
-        assert!(is_keyword_query("embed_batch"));
-        assert!(!is_keyword_query("query"));
-    }
-
-    #[test]
-    fn keyword_natural_language_not_keyword() {
-        assert!(!is_keyword_query("what framework are we using"));
-        assert!(!is_keyword_query("database connection pool"));
-        assert!(!is_keyword_query(""));
-    }
-
-    // ── classify_query_intent tests ─────────────────────────────────────
-
-    #[test]
-    fn intent_keyword_backtick() {
-        assert_eq!(classify_query_intent("`FooBar`"), QueryIntent::Keyword);
-    }
-
-    #[test]
-    fn intent_keyword_file_path() {
-        assert_eq!(classify_query_intent("src/main.rs"), QueryIntent::Keyword);
-    }
-
-    #[test]
-    fn intent_keyword_camel_case() {
-        assert_eq!(classify_query_intent("SqliteStorage"), QueryIntent::Keyword);
-    }
-
-    #[test]
-    fn intent_factual_who() {
-        assert_eq!(
-            classify_query_intent("Who created the memory system?"),
-            QueryIntent::Factual
-        );
-    }
-
-    #[test]
-    fn intent_factual_what() {
-        assert_eq!(
-            classify_query_intent("What database does the project use?"),
-            QueryIntent::Factual
-        );
-    }
-
-    #[test]
-    fn intent_factual_when() {
-        assert_eq!(
-            classify_query_intent("When did we deploy the fix?"),
-            QueryIntent::Factual
-        );
-    }
-
-    #[test]
-    fn intent_factual_where() {
-        assert_eq!(
-            classify_query_intent("Where is the config file stored?"),
-            QueryIntent::Factual
-        );
-    }
-
-    #[test]
-    fn intent_factual_how_many() {
-        assert_eq!(
-            classify_query_intent("How many times has she visited?"),
-            QueryIntent::Factual
-        );
-    }
-
-    #[test]
-    fn intent_factual_how_long() {
-        assert_eq!(
-            classify_query_intent("How long has the project been running?"),
-            QueryIntent::Factual
-        );
-    }
-
-    #[test]
-    fn intent_factual_how_often() {
-        assert_eq!(
-            classify_query_intent("How often does she exercise?"),
-            QueryIntent::Factual
-        );
-    }
-
-    #[test]
-    fn intent_conceptual_why() {
-        assert_eq!(
-            classify_query_intent("Why did we choose SQLite?"),
-            QueryIntent::Conceptual
-        );
-    }
-
-    #[test]
-    fn intent_conceptual_how() {
-        assert_eq!(
-            classify_query_intent("How does the scoring pipeline work?"),
-            QueryIntent::Conceptual
-        );
-    }
-
-    #[test]
-    fn intent_conceptual_explain() {
-        assert_eq!(
-            classify_query_intent("Explain the RRF fusion strategy"),
-            QueryIntent::Conceptual
-        );
-    }
-
-    #[test]
-    fn intent_conceptual_difference() {
-        assert_eq!(
-            classify_query_intent("What is the difference between FTS and vector search?"),
-            QueryIntent::Conceptual
-        );
-    }
-
-    #[test]
-    fn intent_conceptual_compared_to() {
-        assert_eq!(
-            classify_query_intent("performance of SQLite compared to Postgres"),
-            QueryIntent::Conceptual
-        );
-    }
-
-    #[test]
-    fn intent_general_statement() {
-        assert_eq!(
-            classify_query_intent("database connection pool"),
-            QueryIntent::General
-        );
-    }
-
-    #[test]
-    fn intent_general_would() {
-        assert_eq!(
-            classify_query_intent("Would she want to pursue that career?"),
-            QueryIntent::General
-        );
-    }
-
-    #[test]
-    fn intent_general_empty() {
-        assert_eq!(classify_query_intent(""), QueryIntent::General);
-    }
-
-    // ── IntentProfile tests ──────────────────────────────────────────────
-
-    #[test]
-    fn intent_profile_keyword_disables_vec() {
-        let profile = IntentProfile::for_intent(QueryIntent::Keyword);
-        assert!((profile.vec_weight_mult - 0.0).abs() < 1e-9);
-    }
-
-    #[test]
-    fn intent_profile_factual_boosts_fts() {
-        let profile = IntentProfile::for_intent(QueryIntent::Factual);
-        assert!(profile.fts_weight_mult > 1.0);
-        assert!(profile.word_overlap_mult > 1.0);
-    }
-
-    #[test]
-    fn intent_profile_conceptual_boosts_vec() {
-        let profile = IntentProfile::for_intent(QueryIntent::Conceptual);
-        assert!(profile.vec_weight_mult > 1.0);
-        assert!(profile.fts_weight_mult < 1.0);
-    }
-
-    #[test]
-    fn intent_profile_general_is_neutral() {
-        let profile = IntentProfile::for_intent(QueryIntent::General);
-        assert!((profile.vec_weight_mult - 1.0).abs() < 1e-9);
-        assert!((profile.fts_weight_mult - 1.0).abs() < 1e-9);
-        assert!((profile.word_overlap_mult - 1.0).abs() < 1e-9);
-        assert!((profile.top_k_mult - 1.0).abs() < 1e-9);
-    }
-
-    // ── is_lock_error / retry_on_lock tests ────────────────────────────
-
-    #[test]
-    fn is_lock_error_detects_busy() {
-        let err = rusqlite::Error::SqliteFailure(
-            rusqlite::ffi::Error {
-                code: rusqlite::ffi::ErrorCode::DatabaseBusy,
-                extended_code: 5,
-            },
-            Some("database is locked".to_string()),
-        );
-        assert!(is_lock_error(&err));
-    }
-
-    #[test]
-    fn is_lock_error_rejects_other_errors() {
-        let err = rusqlite::Error::SqliteFailure(
-            rusqlite::ffi::Error {
-                code: rusqlite::ffi::ErrorCode::ReadOnly,
-                extended_code: 8,
-            },
-            None,
-        );
-        assert!(!is_lock_error(&err));
-
-        let err2 = rusqlite::Error::QueryReturnedNoRows;
-        assert!(!is_lock_error(&err2));
-    }
-
-    #[test]
-    fn retry_on_lock_returns_immediately_for_non_lock_error() {
-        let mut attempts = 0u32;
-        let result: std::result::Result<(), rusqlite::Error> = retry_on_lock(|| {
-            attempts += 1;
-            Err(rusqlite::Error::QueryReturnedNoRows)
-        });
-        assert!(result.is_err());
-        assert_eq!(attempts, 1, "should not retry for non-lock errors");
-    }
-
-    #[test]
-    fn retry_on_lock_succeeds_on_first_try() {
-        let mut attempts = 0u32;
-        let result = retry_on_lock(|| {
-            attempts += 1;
-            Ok(42)
-        });
-        assert_eq!(result.unwrap(), 42);
-        assert_eq!(attempts, 1);
-    }
-
-    #[test]
-    fn retry_on_lock_retries_on_busy_then_succeeds() {
-        let mut attempts = 0u32;
-        let result = retry_on_lock(|| {
-            attempts += 1;
-            if attempts < 3 {
-                Err(rusqlite::Error::SqliteFailure(
-                    rusqlite::ffi::Error {
-                        code: rusqlite::ffi::ErrorCode::DatabaseBusy,
-                        extended_code: 5,
-                    },
-                    None,
-                ))
-            } else {
-                Ok("success")
-            }
-        });
-        assert_eq!(result.unwrap(), "success");
-        assert_eq!(attempts, 3);
-    }
-
-    #[test]
-    fn retry_on_lock_exhausts_retries() {
-        let mut attempts = 0u32;
-        let result: std::result::Result<(), rusqlite::Error> = retry_on_lock(|| {
-            attempts += 1;
-            Err(rusqlite::Error::SqliteFailure(
-                rusqlite::ffi::Error {
-                    code: rusqlite::ffi::ErrorCode::DatabaseBusy,
-                    extended_code: 5,
-                },
-                None,
-            ))
-        });
-        assert!(result.is_err());
-        assert_eq!(
-            attempts, 5,
-            "should attempt exactly RETRY_MAX_ATTEMPTS times"
         );
     }
 
@@ -2778,107 +1233,5 @@ mod tests {
         assert!(q.contains("\"movie\""));
         assert!(q.contains("\"film\""));
         assert!(!q.contains("\"the\""));
-    }
-
-    #[test]
-    fn test_dynamic_limit_mult_multi_hop() {
-        assert!(
-            (detect_dynamic_limit_mult("What is Amanda's sister and how are they related") - 2.0)
-                .abs()
-                < 1e-9
-        );
-        assert!(
-            (detect_dynamic_limit_mult("Tell me about the relationship between Alice and Bob")
-                - 2.0)
-                .abs()
-                < 1e-9
-        );
-    }
-
-    #[test]
-    fn test_dynamic_limit_mult_temporal() {
-        assert!((detect_dynamic_limit_mult("What happened last month") - 1.5).abs() < 1e-9);
-        assert!((detect_dynamic_limit_mult("Events over the past year") - 1.5).abs() < 1e-9);
-    }
-
-    #[test]
-    fn test_dynamic_limit_mult_simple() {
-        assert!((detect_dynamic_limit_mult("What is Alice's job?") - 1.0).abs() < 1e-9);
-        assert!((detect_dynamic_limit_mult("Tell me about the weather") - 1.0).abs() < 1e-9);
-    }
-
-    #[test]
-    fn test_intent_profile_suggested_limit_mult() {
-        let p = IntentProfile::for_intent(QueryIntent::Conceptual);
-        assert!((p.suggested_limit_mult - 1.3).abs() < 1e-9);
-        let p = IntentProfile::for_intent(QueryIntent::General);
-        assert!((p.suggested_limit_mult - 1.0).abs() < 1e-9);
-    }
-
-    // ── extract_query_entities tests ───────────────────────────────────
-
-    #[test]
-    fn test_extract_query_entities() {
-        let entities = extract_query_entities("Would Caroline pursue writing?");
-        assert!(
-            entities.contains(&"Caroline".to_string()),
-            "expected Caroline in {entities:?}"
-        );
-        assert!(
-            !entities.contains(&"Would".to_string()),
-            "should skip question word Would"
-        );
-    }
-
-    #[test]
-    fn test_extract_query_entities_possessive() {
-        let entities = extract_query_entities("What is Amanda's sister's career?");
-        assert!(
-            entities.contains(&"Amanda".to_string()),
-            "expected Amanda in {entities:?}"
-        );
-    }
-
-    // ── extract_topic_keywords tests ───────────────────────────────────
-
-    #[test]
-    fn test_extract_topic_keywords() {
-        let topics = extract_topic_keywords(
-            "Would Caroline pursue writing as a career?",
-            &["Caroline".to_string()],
-        );
-        assert!(
-            topics.contains(&"writing".to_string()),
-            "expected 'writing' in {topics:?}"
-        );
-        assert!(
-            topics.contains(&"career".to_string()),
-            "expected 'career' in {topics:?}"
-        );
-        assert!(
-            !topics.contains(&"caroline".to_string()),
-            "should exclude entity 'caroline'"
-        );
-    }
-
-    // ── generate_sub_queries tests ─────────────────────────────────────
-
-    #[test]
-    fn test_generate_sub_queries() {
-        let queries = generate_sub_queries(
-            "Would Caroline pursue writing?",
-            &["Caroline".to_string()],
-            &["writing".to_string()],
-        );
-        assert_eq!(queries[0], "Would Caroline pursue writing?");
-        assert!(queries.contains(&"Caroline".to_string()));
-        assert!(queries.contains(&"Caroline writing".to_string()));
-    }
-
-    #[test]
-    fn test_generate_sub_queries_no_entities() {
-        let queries = generate_sub_queries("Tell me about writing", &[], &["writing".to_string()]);
-        assert_eq!(queries[0], "Tell me about writing");
-        assert!(queries.contains(&"writing".to_string()));
     }
 }
